feat(core): group the whole client surface by protocol category
Requests and conveniences now sit together in the category they belong to,
so client.Scenes.GetSceneListAsync and client.Scenes.SwitchProgramSceneAsync
read the same way and IntelliSense offers 14 categories rather than 147
flat methods. The groups are partial structs, so generated and hand-written
members share one type and same-named overloads resolve naturally.

The flat extension methods are gone; the batch builder already used these
categories, so both halves of the library now address OBS identically.

SceneItemNotFoundException moves out of the static helper class it was
nested in.
